+++ src/com/intellij/idea/plugin/hybris/common/HybrisConstants.kt
@@ -182,6 +182,9 @@ object HybrisConstants {
     const val PROPERTY_OPTIONAL_CONFIG_DIR = "hybris.optional.config.dir"
     const val PROPERTY_LANG_PACKS = "lang.packs"
     const val PROPERTY_IMPEX_HEADER_REPLACEMENT = "impex.header.replacement"
+    const val PROPERTY_ENV_PROPERTY_PREFIX = "env.properties.prefix"
+
+    const val PROPERTY_STANDALONE_JDKMODULESEXPORTS = "standalone.jdkmodulesexports";
 
     const val DEFAULT_LANGUAGE_ISOCODE = "en"
 
@@ -301,13 +304,17 @@ object HybrisConstants {
     const val LIB_DIRECTORY = "lib"
     const val BIN_DIRECTORY = "bin"
     const val RESOURCES_DIRECTORY = "resources"
-    const val LOCAL_PROPERTIES = "local.properties"
-    const val PROJECT_PROPERTIES = "project.properties"
+    const val LOCAL_PROPERTIES_FILE = "local.properties"
+    const val PROJECT_PROPERTIES_FILE = "project.properties"
+    const val ENV_PROPERTIES_FILE = "env.properties"
+    const val ADVANCED_PROPERTIES_FILE = "advanced.properties"
     const val SPRING_WEB_FILE_SET_NAME = "web application context"
     const val APPLICATION_CONTEXT_SPRING_FILES = "application-context"
     const val ADDITIONAL_WEB_SPRING_CONFIG_FILES = "additionalWebSpringConfigs"
     const val GLOBAL_CONTEXT_SPRING_FILES = "global-context"
-    const val HYBRIS_CONFIG_DIR_KEY = "HYBRIS_CONFIG_DIR"
+    const val HYBRIS_CONFIG_DIR_ENV = "HYBRIS_CONFIG_DIR"
+    const val HYBRIS_RUNTIME_PROPERTIES_ENV = "HYBRIS_RUNTIME_PROPERTIES"
+    const val HYBRIS_OPT_CONFIG_DIR_ENV = "HYBRIS_OPT_CONFIG_DIR"
     const val HYBRIS_API_VERSION_KEY = "version.api"
     const val HYBRIS_VERSION_KEY = "version"
     const val JAVADOC_FALLBACK_URL = "https://help.sap.com/docs/SAP_COMMERCE/c5613bd3cc9942efb74d017b40eb0892/179bbc9b35274d7ca784e46b3beb40b2.html"