+++ src/com/intellij/idea/plugin/hybris/runConfigurations/HybrisJUnitExtension.java
@@ -18,6 +18,8 @@
 
 package com.intellij.idea.plugin.hybris.runConfigurations;
 
+import static com.intellij.idea.plugin.hybris.common.HybrisConstants.*;
+
 import com.intellij.execution.ExecutionException;
 import com.intellij.execution.RunConfigurationExtension;
 import com.intellij.execution.configurations.JavaParameters;
@@ -26,47 +28,73 @@ import com.intellij.execution.configurations.RunConfigurationBase;
 import com.intellij.execution.configurations.RunnerSettings;
 import com.intellij.execution.junit.JUnitConfiguration;
 import com.intellij.idea.plugin.hybris.project.utils.HybrisRootUtil;
+import com.intellij.idea.plugin.hybris.properties.PropertiesService;
 import com.intellij.idea.plugin.hybris.settings.HybrisProjectSettings;
 import com.intellij.idea.plugin.hybris.settings.HybrisProjectSettingsComponent;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.io.FileUtil;
 import com.intellij.openapi.vfs.VirtualFile;
-import org.jetbrains.annotations.NotNull;
 
-import static com.intellij.idea.plugin.hybris.common.HybrisConstants.HYBRIS_DATA_DIRECTORY;
-import static com.intellij.idea.plugin.hybris.common.HybrisConstants.HYBRIS_DATA_DIR_ENV;
+import java.util.StringTokenizer;
+
+import org.jetbrains.annotations.NotNull;
 
 public class HybrisJUnitExtension extends RunConfigurationExtension {
 
     @Override
     public <T extends RunConfigurationBase<?>> void updateJavaParameters(
-        final T configuration, final JavaParameters params, final RunnerSettings runnerSettings
-    ) throws ExecutionException {
+        final T configuration, final JavaParameters params, final RunnerSettings runnerSettings)
+        throws ExecutionException {
         if (runnerSettings != null || !isApplicableFor(configuration)) {
             return;
         }
         final Project project = configuration.getProject();
         final ParametersList vmParameters = params.getVMParametersList();
 
-        if (!vmParameters.hasParameter("-ea")) {
-            vmParameters.add("-ea");
-        }
-        if (vmParameters.getParameters().stream().noneMatch(param -> param.startsWith("-Dplatformhome="))) {
+        addVmParameterIfNotExist(vmParameters, "-ea");
+
+        if (vmParameters.getParameters().stream()
+            .noneMatch(param -> param.startsWith("-Dplatformhome="))) {
             final VirtualFile platformRootDirectory = HybrisRootUtil.findPlatformRootDirectory(project);
 
             if (platformRootDirectory != null) {
                 vmParameters.add("-Dplatformhome=" + platformRootDirectory.getPath());
             }
         }
-        if (!params.getEnv().containsKey(HYBRIS_DATA_DIR_ENV)) {
-            final HybrisProjectSettings settings = HybrisProjectSettingsComponent.getInstance(project).getState();
 
-            final String hybrisDataDirPath = FileUtil.toCanonicalPath(
-                project.getBasePath() + '/' + settings.getHybrisDirectory() + '/' + HYBRIS_DATA_DIRECTORY);
+        if (!params.getEnv().containsKey(HYBRIS_DATA_DIR_ENV)) {
+            final HybrisProjectSettings settings =
+                HybrisProjectSettingsComponent.getInstance(project).getState();
+
+            final String hybrisDataDirPath =
+                FileUtil.toCanonicalPath(
+                    project.getBasePath()
+                        + '/'
+                        + settings.getHybrisDirectory()
+                        + '/'
+                        + HYBRIS_DATA_DIRECTORY);
 
             if (hybrisDataDirPath != null) {
                 params.addEnv(HYBRIS_DATA_DIR_ENV, hybrisDataDirPath);
             }
+        }
+
+        final PropertiesService propertiesService = PropertiesService.getInstance(project);
+        if (propertiesService != null) {
+            final String property = propertiesService.findProperty(PROPERTY_STANDALONE_JDKMODULESEXPORTS);
+            if (property != null) {
+                final StringTokenizer tokenizer = new StringTokenizer(property.trim());
+                while (tokenizer.hasMoreTokens()) {
+                    final String newParam = tokenizer.nextToken().replaceAll("\"", "");
+                    addVmParameterIfNotExist(vmParameters, newParam);
+                }
+            }
+        }
+    }
+
+    private void addVmParameterIfNotExist(ParametersList vmParameters, String newParam) {
+        if (!vmParameters.hasParameter(newParam)) {
+            vmParameters.add(newParam);
         }
     }
 
@@ -78,5 +106,4 @@ public class HybrisJUnitExtension extends RunConfigurationExtension {
         final Project project = configuration.getProject();
         return HybrisProjectSettingsComponent.getInstance(project).isHybrisProject();
     }
-
 }