+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [2023.3.3]
+
+### `Tests` enhancements
+- Add JDK Export Properties to JUnit Tests to run Integration Tests [#864](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/864)
+- Load Properties from Files and Environment [#864](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/864)
+
 ## [2023.3.2]
 
 ### `Project Import` enhancements